+++ src/main/java/co/kr/jurumarble/config/WebConfig.java
@@ -1,13 +1,28 @@
 package co.kr.jurumarble.config;
 
+import co.kr.jurumarble.interceptor.JwtInterceptor;
+import co.kr.jurumarble.token.domain.JwtTokenProvider;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
+@RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
 
     public static final String ALLOWED_METHOD_NAMES = "GET,HEAD,POST,PUT,DELETE,TRACE,OPTIONS,PATCH";
+    private final JwtInterceptor jwtInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(jwtInterceptor)
+                .order(1)
+                .addPathPatterns("/api/users/additional-info");
+    }
 
     @Override
     public void addCorsMappings(final CorsRegistry registry) {
