--- src/main/java/co/kr/jurumarble/config/WebFilterConfig.java
@@ -1,21 +1,0 @@
-package co.kr.jurumarble.config;
-
-import co.kr.jurumarble.filter.JwtAuthenticationFilter;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import javax.servlet.Filter;
-
-@Configuration
-public class WebFilterConfig {
-
-    private static final String[] filterUrls = {"/api/users/additional-info", "/api/users/additional-category"};
-
-    @Bean
-    public FilterRegistrationBean JwtFilter(JwtAuthenticationFilter jwtAuthenticationFilter) {
-        FilterRegistrationBean<Filter> filterRegistrationBean = new FilterRegistrationBean<>(jwtAuthenticationFilter);
-        filterRegistrationBean.addUrlPatterns(filterUrls);
-        return filterRegistrationBean;
-    }
-}
